+++ tests/szconfig_test.py
@@ -6,7 +6,7 @@ from typing import Any, Dict
 
 import pytest
 from pytest_schema import Optional, Or, schema
-from senzing import SzConfig, SzSdkError
+from senzing import SzConfig, SzError, SzSdkError
 
 from senzing_core import SzConfigCore
 
@@ -38,6 +38,13 @@ def test_add_data_source_bad_data_source_code_value(sz_config: SzConfig) -> None
         sz_config.add_data_source(bad_data_source_code)  # type: ignore[arg-type]
 
 
+def test_add_data_source_empty_data_source_code_value(sz_config: SzConfig) -> None:
+    """Test SzConfig.add_data_source()."""
+    bad_data_source_code = ""
+    with pytest.raises(SzError):
+        sz_config.add_data_source(bad_data_source_code)
+
+
 def test_delete_data_source(sz_config: SzConfig) -> None:
     """Test SzConfig.delete_data_source()."""
     data_source_code = "TEST"
@@ -56,6 +63,13 @@ def test_delete_data_source_bad_data_source_code_value(sz_config: SzConfig) -> N
     bad_data_source_code = {"XXXX": "YYYY"}
     with pytest.raises(SzSdkError):
         sz_config.delete_data_source(bad_data_source_code)  # type: ignore[arg-type]
+
+
+def test_delete_data_source_empty_data_source_code_value(sz_config: SzConfig) -> None:
+    """Test SzConfig.delete_data_source()."""
+    bad_data_source_code = ""
+    with pytest.raises(SzError):
+        sz_config.delete_data_source(bad_data_source_code)
 
 
 def test_export(sz_config: SzConfig) -> None: